Create README.md
I'm a highly motivated and inspired by completion individual. My life has always been about " earning the quick buck" and until recently I had never thought about a career development in my life... well.. now I have a child and a wife and I now think it is time to learn something that will not only create a great life for them but to also be able to learn skills that I can pass on to my child so that he can have some core skills in a fantastic trade. I am open to whoever would like to work on something or can offer advice, my door is always an open book!